# Implement lazy funcref table and anyfunc initialization.