@cfallin 2022-01-27
Benchmarking: the delta is hard to see with memory initialization taking most of the time (pre-#3699); a few percent improvement in the mean with `spidermonkey.wasm` but Criterion says not statistically significant.

However, in #3697 I [measured](https://github.com/bytecodealliance/wasmtime/pull/3697#issuecomment-1020711683) approximately a factor-of-two improvement in the `initialize_vmcontext` hotpath (which was most of instantiation, in a scenario where we have a memory technique to reuse mappings).