@alexcrichton 2022-01-27
Thanks for splitting this out! I've mentioned a few things here in other contexts but I think it's good to get them all in once place. The tl;dr; is that I don't personally feel like this is the best approach, but I think we can still get the same wins.

Due to a number of implementation details I don't think that this PR is really getting all that much of a benefit from laziness. The anyfunc array is used as storage for a module's `*mut VMCallerCheckedAnyfunc` which is used as the representation of a `funcref`. This is subsequently used by things like `ref.func`, tables, and exports. Today the anyfunc array is the size of the module's index space of functions, but this is actually much larger than necessary because we only use it for those items, and not all functions make their way into `ref.func`, tables, or exports. This PR also doesn't lazily initialize tables, only the anyfunc array, which means we still initialize all anyfuncs used as part of tables.

Adding all that together we still pay the full cost of anyfunc initialization for all table entries when a module is instantiated. The cost of initializing exports is done lazily and most modules nowadays don't have `ref.func` instructions so there's not much cost there. Overall this means that this PR is improving the amount of work done, but I think it's doing it in a bit of a roundabout way. Effectively we're immediately initializing a static subset of the anyfunc array on initialization and we're never actually initializing anything else after that (modulo exports but that's such a small set I don't think it really matters all that much).

Given all that personally think that a better implementation would be to avoid the laziness complications and instead just go ahead and keep doing what we do today, only over the same set of functions that this PR is doing it for. Basically I think we should shrink the anyfunc array to the size of [the `escaped_funcs` array](https://github.com/bytecodealliance/wasmtime/blob/7928a3ffb4ddce1d60ef7634a9cf79c68258c1c0/crates/environ/src/module_environ.rs#L64-L74). That I think means that the same magnitude of work would be done on instantiation and there wouldn't be any need to deal with laziness anywhere.

---

As a few other side things:

* For "benchmarking" this I think it might be good to take a module with 1k functions and a table with 100 of those functions and instantiate that. That should get memory out of the picture and should help focus on just the time it takes to deal with all the anyfuncs.
* For the race comment, I'm under the impression that because the writes are non-atomic that any concurrent access to `get_caller_checked_anyfunc` would result in a data race. I don't think we have concurrent accesses today though so this is probably fine, but if we were to stick with this PR I'd prefer to either make `get_caller_checked_anyfunc` unsafe or make it take `&mut self`.

---

Ok now all that being said, one thing you mentioned the other day which I'm coming around to is that long-term I think it would be best to lazily initialize tables instead of just this anyfunc array. With CoW and uffd we're already effectively lazily initializing memory (just relying on the kernel to do so), and I think it would be best to do that for tables as well. I don't really know how we'd do this off the top of my head though.

Alternatively we could try to figure out how to make the table initialization much more "static" where work is done at compile time instead of instantiation time. I had one idea [over here](https://github.com/bytecodealliance/wasmtime/pull/3697#issuecomment-1020240918) but that's sort of half-baked and only shrinks `VMCallerCheckedAnyfunc`, it doesn't offload too much to compile time. Overall doing something at compile-time I think would also be better than the lazy approach when we can.