@cfallin 2022-02-03
I've implemented lazy table initialization now (mostly; there's one odd test failure with linking I need to resolve). It could probably be cleaned up somewhat, but here's one initial datapoint, with spidermonkey.wasm:

```
(before)

sequential/default/spidermonkey.wasm
                        time:   [71.886 us 72.012 us 72.133 us]

(after)

sequential/default/spidermonkey.wasm
                        time:   [22.243 us 22.256 us 22.270 us]
                        change: [-69.117% -69.060% -69.000%] (p = 0.00 < 0.05)
                        Performance has improved.
```

So, 72µs to 22µs, or a 69% reduction.